async_destroy_lambda/requirements.txt
gitpython
wget
zipfile